requirements.txt
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.0
attrs==24.2.0
Authlib==1.3.2
backoff==2.2.1
beautifulsoup4==4.12.3
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
ConfigArgParse==1.7
cryptography==43.0.3
diff-match-patch==20230430
diskcache==5.6.3
distro==1.9.0
fastapi==0.115.2
filelock==3.16.1
flake8==7.1.1
frozenlist==1.4.1
fsspec==2024.9.0
gitdb==4.0.11
GitPython==3.1.43
grep-ast==0.3.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.25.0
idna==3.10
importlib_metadata==7.2.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.5.0
json5==0.9.25
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
litellm==1.47.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
multidict==6.1.0
networkx==3.2.1
numpy==1.26.4
oauthlib==3.2.2
openai==1.47.0
packaging==24.1
pathspec==0.12.1
pexpect==4.9.0
pillow==10.4.0
prompt_toolkit==3.0.47
psutil==6.0.0
ptyprocess==0.7.0
pycodestyle==2.12.1
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
pydub==0.25.1
pyflakes==3.2.0
Pygments==2.18.0
pypandoc==1.13
pyperclip==1.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.16
PyYAML==6.0.2
referencing==0.35.1
regex==2024.9.11
requests==2.32.3
requests-oauthlib==2.0.0
rich==13.8.1
rpds-py==0.20.0
scipy==1.13.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.1
sounddevice==0.5.0
soundfile==0.12.1
soupsieve==2.6
starlette==0.40.0
tiktoken==0.7.0
tokenizers==0.19.1
tqdm==4.66.5
tree-sitter==0.21.3
tree-sitter-languages==1.10.2
typing_extensions==4.12.2
urllib3==2.2.3
uvicorn==0.32.0
wcwidth==0.2.13
xero_python==6.3.0
yarl==1.11.1
zipp==3.20.2
